# Conflicts:
#	frontend/src/components/wrapper.jsx
#	routes/web.php

+++ database/migrations/2018_07_28_123531_add_yes_or_maybe_field_to_can_sing.php
@@ -25,8 +25,11 @@ class AddYesOrMaybeFieldToCanSing extends Migration
      */
     public function down()
     {
-        Schema::table('can_sing', function (Blueprint $table) {
-            $table->dropColumn('yes_or_maybe');
-        });
+        if (Schema::hasColumn('can_sing', 'yes_or_maybe'))
+        {
+            Schema::table('can_sing', function (Blueprint $table) {
+                $table->dropColumn('yes_or_maybe');
+            });
+        }
     }
 }
